change `reset-device-id` to subcommand.
